requirements.txt
pyyaml>=6.0
numpy>=1.23
tqdm>=4.66
torch>=2.0
